refactor: error 클래스 경로 수정
