+++ src/auth/service.ts
@@ -2,7 +2,8 @@ import bcrypt from 'bcrypt';
 import { AuthUserPayload } from './dto/login.dto';
 import AuthRepository from './repository';
 import { signAccessToken, signRefreshToken, verifyRefreshToken, RefreshTokenPayload } from './jwt';
-import { NotFoundError, UnauthorizedError } from '../middlewares/error.middleware';
+import { NotFoundError } from '../common/errors/not-found-error';
+import { UnauthorizedError } from '../common/errors/unauthorize-error';
 
 export interface LoginResponse {
   user: {
